+++ i18n.sh
@@ -30,7 +30,7 @@ log_message "Package version: $PACKAGE_VERSION"
 
 # Generate pot file from .js files, excluding node_modules directory
 log_message "Generating POT file..."
-JS_FILES=$(find . -name '*.js' -type f -not -path './node_modules/*')
+JS_FILES=$(find . -name '*.js' -type f -not -path './node_modules/*' -not -path './dist/*')
 xgettext --language=JavaScript --from-code=UTF-8 --package-name="$PACKAGE_NAME" --package-version="$PACKAGE_VERSION" --copyright-holder="Lju" --output=po/monitor@astraext.pot $JS_FILES
 if [ $? -ne 0 ]; then
     log_message "Error: Failed to generate POT file"