+++ RELEASES.md
@@ -1,8 +1,13 @@
 # Astra Monitor 7 - TO BE RELEASED
 
 ### New features
+- GTop support as _data source_ for Top Processes is now available in CPU Menu [[#33](https://github.com/AstraExt/astra-monitor/issues/33)]
+- New option to show cpu percentage on Top Processes list as per-core value [[#33](https://github.com/AstraExt/astra-monitor/issues/33)]
 - Regex Ignore Network Interfaces: now, in addition to manual selection, you can use regex to ignore network interfaces [[#29](https://github.com/AstraExt/astra-monitor/issues/29)]
 - New languages: German and Russian [[#26](https://github.com/AstraExt/astra-monitor/issues/26)] [[#27](https://github.com/AstraExt/astra-monitor/issues/27)]
+
+### Bug fixes
+- Top Processes list in Processors Menu is now even more responsive and should be filled faster for both '_GTop_' and '_/proc_' based sources
 
 # Astra Monitor 6 - January 19 2024